+++ src/lib/util.js
@@ -2,7 +2,7 @@
 
 import { CARD } from "@paypal/sdk-constants/src";
 
-import { CDN_BASE_URL, LOGO, LOGO_COLOR, MARK } from "../constants";
+import { CDN_BASE_URL, LOGO, LOGO_COLOR, MARK, BADGE } from "../constants";
 import type { LogoColorMap, LogoColors } from "../types";
 
 export function getLogoColors(
@@ -30,7 +30,11 @@ export function getLogoColors(
 }
 
 export function getSVGFilename(
-  logoName: $Values<typeof LOGO> | $Values<typeof CARD> | $Values<typeof MARK>,
+  logoName:
+    | $Values<typeof LOGO>
+    | $Values<typeof CARD>
+    | $Values<typeof MARK>
+    | $Values<typeof BADGE>,
   logoColor?: $Values<typeof LOGO_COLOR>
 ): string {
   if (logoColor) {
@@ -41,7 +45,11 @@ export function getSVGFilename(
 }
 
 export function getLogoCDNUrl(
-  logoName: $Values<typeof LOGO> | $Values<typeof CARD> | $Values<typeof MARK>,
+  logoName:
+    | $Values<typeof LOGO>
+    | $Values<typeof CARD>
+    | $Values<typeof MARK>
+    | $Values<typeof BADGE>,
   logoColorMap?: LogoColorMap,
   logoColor?: $Values<typeof LOGO_COLOR>
 ): string {
