+++ src/logos/creditRebrand/index.js
@@ -2,3 +2,4 @@
 
 export * from "./mark";
 export * from "./logo";
+export * from "./badge";
