+++ src/logos/paypalRebrand/index.js
@@ -1,3 +1,4 @@
 /* @flow */
 
 export * from "./logo";
+export * from "./badge";
